Controles de acceso y sessiones

+++ regiones/delete.php
@@ -4,40 +4,47 @@ require('../class/rutas.php');
 
 session_start();
 
-if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
-    $id = (int) $_POST['id'];
+if(isset($_SESSION['autenticado']) && $_SESSION['usuario_rol'] == 'Administrador'){
 
-    #verificar que hay una region con el id enviado
-    $res = $mbd->prepare("SELECT id FROM regiones WHERE id = ?");
-    $res->bindParam(1, $id);
-    $res->execute();
-    $region = $res->fetch();
 
-    if ($region) {
-        #verificar que la region existente no tenga comunas asociadas
-        $res = $mbd->prepare("SELECT id FROM comunas WHERE region_id = ?");
+    if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
+        $id = (int) $_POST['id'];
+
+        #verificar que hay una region con el id enviado
+        $res = $mbd->prepare("SELECT id FROM regiones WHERE id = ?");
         $res->bindParam(1, $id);
         $res->execute();
-        $comunas = $res->fetchall();
+        $region = $res->fetch();
 
-        if (!$comunas) {
-            #eliminar la region
-            $res = $mbd->prepare("DELETE FROM regiones WHERE id = ?");
+        if ($region) {
+            #verificar que la region existente no tenga comunas asociadas
+            $res = $mbd->prepare("SELECT id FROM comunas WHERE region_id = ?");
             $res->bindParam(1, $id);
             $res->execute();
+            $comunas = $res->fetchall();
 
-            $row = $res->rowCount();
+            if (!$comunas) {
+                #eliminar la region
+                $res = $mbd->prepare("DELETE FROM regiones WHERE id = ?");
+                $res->bindParam(1, $id);
+                $res->execute();
 
-            if ($row) {
-                $_SESSION['success'] = 'La region se ha eliminado correctamente';
-                header('Location: ' . REGIONES);
+                $row = $res->rowCount();
+
+                if ($row) {
+                    $_SESSION['success'] = 'La region se ha eliminado correctamente';
+                    header('Location: ' . REGIONES);
+                }
+            }else{
+                $_SESSION['danger'] = 'La región tiene comunas asociadas... no se puede eliminar';
+                header('Location: ' . REGIONES . 'show.php?id=' . $id);
             }
-        }else{
-            $_SESSION['danger'] = 'La región tiene comunas asociadas... no se puede eliminar';
+        }else {
+            $_SESSION['danger'] = 'La región no se puede eliminar';
             header('Location: ' . REGIONES . 'show.php?id=' . $id);
         }
-    }else {
-        $_SESSION['danger'] = 'La región no se puede eliminar';
-        header('Location: ' . REGIONES . 'show.php?id=' . $id);
     }
+}else{
+    $_SESSION['danger'] = 'Operación no permitida';
+    header('Location: ' . BASE_URL);
 }
